refactor: split engine/assignment/strategies.py into per-strategy modules
