+++ src/synthorg/engine/assignment/__init__.py
@@ -5,6 +5,19 @@ designation, role-based scoring, load-balanced selection,
 cost-optimized selection, hierarchical delegation, or auction.
 """
 
+from synthorg.engine.assignment._shared import (
+    STRATEGY_NAME_AUCTION,
+    STRATEGY_NAME_COST_OPTIMIZED,
+    STRATEGY_NAME_HIERARCHICAL,
+    STRATEGY_NAME_LOAD_BALANCED,
+    STRATEGY_NAME_MANUAL,
+    STRATEGY_NAME_ROLE_BASED,
+)
+from synthorg.engine.assignment.auction import AuctionAssignmentStrategy
+from synthorg.engine.assignment.cost_optimized import CostOptimizedAssignmentStrategy
+from synthorg.engine.assignment.hierarchical import HierarchicalAssignmentStrategy
+from synthorg.engine.assignment.load_balanced import LoadBalancedAssignmentStrategy
+from synthorg.engine.assignment.manual import ManualAssignmentStrategy
 from synthorg.engine.assignment.models import (
     AgentWorkload,
     AssignmentCandidate,
@@ -16,21 +29,8 @@ from synthorg.engine.assignment.registry import (
     STRATEGY_MAP,
     build_strategy_map,
 )
+from synthorg.engine.assignment.role_based import RoleBasedAssignmentStrategy
 from synthorg.engine.assignment.service import TaskAssignmentService
-from synthorg.engine.assignment.strategies import (
-    STRATEGY_NAME_AUCTION,
-    STRATEGY_NAME_COST_OPTIMIZED,
-    STRATEGY_NAME_HIERARCHICAL,
-    STRATEGY_NAME_LOAD_BALANCED,
-    STRATEGY_NAME_MANUAL,
-    STRATEGY_NAME_ROLE_BASED,
-    AuctionAssignmentStrategy,
-    CostOptimizedAssignmentStrategy,
-    HierarchicalAssignmentStrategy,
-    LoadBalancedAssignmentStrategy,
-    ManualAssignmentStrategy,
-    RoleBasedAssignmentStrategy,
-)
 
 __all__ = [
     "STRATEGY_MAP",
