+++ tests/unit/engine/test_assignment_project_team.py
@@ -9,8 +9,8 @@ from synthorg.core.enums import SeniorityLevel, TaskType
 from synthorg.core.role import Skill
 from synthorg.core.task import Task
 from synthorg.engine.assignment.models import AssignmentRequest
+from synthorg.engine.assignment.role_based import RoleBasedAssignmentStrategy
 from synthorg.engine.assignment.service import TaskAssignmentService
-from synthorg.engine.assignment.strategies import RoleBasedAssignmentStrategy
 from synthorg.engine.routing.scorer import AgentTaskScorer
 
 pytestmark = pytest.mark.unit
