+++ tests/unit/engine/test_strategy_registry.py
@@ -6,25 +6,25 @@ import pytest
 
 from synthorg.communication.delegation.hierarchy import HierarchyResolver
 from synthorg.core.company import Company, Department, Team
-from synthorg.engine.assignment.protocol import TaskAssignmentStrategy
-from synthorg.engine.assignment.registry import (
-    STRATEGY_MAP,
-    build_strategy_map,
-)
-from synthorg.engine.assignment.strategies import (
+from synthorg.engine.assignment._shared import (
     STRATEGY_NAME_AUCTION,
     STRATEGY_NAME_COST_OPTIMIZED,
     STRATEGY_NAME_HIERARCHICAL,
     STRATEGY_NAME_LOAD_BALANCED,
     STRATEGY_NAME_MANUAL,
     STRATEGY_NAME_ROLE_BASED,
-    AuctionAssignmentStrategy,
-    CostOptimizedAssignmentStrategy,
-    HierarchicalAssignmentStrategy,
-    LoadBalancedAssignmentStrategy,
-    ManualAssignmentStrategy,
-    RoleBasedAssignmentStrategy,
 )
+from synthorg.engine.assignment.auction import AuctionAssignmentStrategy
+from synthorg.engine.assignment.cost_optimized import CostOptimizedAssignmentStrategy
+from synthorg.engine.assignment.hierarchical import HierarchicalAssignmentStrategy
+from synthorg.engine.assignment.load_balanced import LoadBalancedAssignmentStrategy
+from synthorg.engine.assignment.manual import ManualAssignmentStrategy
+from synthorg.engine.assignment.protocol import TaskAssignmentStrategy
+from synthorg.engine.assignment.registry import (
+    STRATEGY_MAP,
+    build_strategy_map,
+)
+from synthorg.engine.assignment.role_based import RoleBasedAssignmentStrategy
 from synthorg.engine.routing.scorer import AgentTaskScorer
 
 pytestmark = pytest.mark.unit
